[skip ci] feat: move related assets checks logic to assertValidTrade
