+++ src/lib/swapper/swappers/ArbitrumBridgeSwapper/getTradeQuote/getTradeQuote.ts
@@ -18,7 +18,6 @@ import { Err, Ok } from '@sniptt/monads'
 import { getDefaultSlippageDecimalPercentageForSwapper } from 'constants/constants'
 import { BigNumber } from 'ethers5'
 import { v4 as uuid } from 'uuid'
-import { getAddress, isAddressEqual } from 'viem'
 import { arbitrum } from 'viem/chains'
 import { usdcAssetId } from 'components/Modals/FiatRamps/config'
 import { bn } from 'lib/bignumber/bignumber'
@@ -50,7 +49,7 @@ export async function getTradeQuote(
     sellAmountIncludingProtocolFeesCryptoBaseUnit,
     sendAddress,
   } = input
-  const assertion = assertValidTrade({ buyAsset, sellAsset })
+  const assertion = await assertValidTrade({ buyAsset, sellAsset })
   if (assertion.isErr()) return Err(assertion.unwrapErr())
 
   const adapter = assertGetEvmChainAdapter(chainId)
@@ -59,43 +58,11 @@ export async function getTradeQuote(
 
   const isDeposit = sellAsset.chainId === ethChainId
   const isEthBridge = isDeposit ? sellAsset.assetId === ethAssetId : buyAsset.assetId === ethAssetId
-  const isTokenBridge = !isEthBridge
 
   const bridger = isEthBridge ? new EthBridger(l2Network) : new Erc20Bridger(l2Network)
   const erc20L1Address = fromAssetId((isDeposit ? sellAsset : buyAsset).assetId).assetReference
-  const erc20L2Address = fromAssetId((isDeposit ? buyAsset : sellAsset).assetId).assetReference
   const l1Provider = getEthersV5Provider(KnownChainIds.EthereumMainnet)
   const l2Provider = getEthersV5Provider(KnownChainIds.ArbitrumMainnet)
-
-  if (isTokenBridge) {
-    // Since our related assets list isn't exhaustive and won't cut it to determine the L1 <-> L2 mapping, we double check that the bridge is valid
-    // by checking against Arbitrum bridge's own mappings, which uses different sources (Coingecko, Gemini, Uni and its own lists at the time of writing)
-    const arbitrumBridgeErc20L2Address = await (bridger as Erc20Bridger).getL2ERC20Address(
-      erc20L1Address,
-      l1Provider,
-    )
-    const arbitrumBridgeErc20L1Address = await (bridger as Erc20Bridger).getL1ERC20Address(
-      erc20L2Address,
-      l2Provider,
-    )
-
-    if (!isAddressEqual(getAddress(arbitrumBridgeErc20L1Address), getAddress(erc20L1Address))) {
-      return Err(
-        makeSwapErrorRight({
-          message: `[ArbitrumBridge: tradeQuote] - Invalid L1 ERC20 address: ${erc20L1Address}`,
-          code: TradeQuoteError.UnsupportedTradePair,
-        }),
-      )
-    }
-    if (!isAddressEqual(getAddress(arbitrumBridgeErc20L2Address), getAddress(erc20L2Address))) {
-      return Err(
-        makeSwapErrorRight({
-          message: `[ArbitrumBridge: tradeQuote] - Invalid L2 ERC20 address: ${erc20L2Address}`,
-          code: TradeQuoteError.UnsupportedTradePair,
-        }),
-      )
-    }
-  }
 
   if (sellAsset.assetId === usdcAssetId || sellAsset.assetId === usdcOnArbitrumAssetId) {
     // https://www.circle.com/en/cross-chain-transfer-protocol
